Refactor into common interface
